requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
djangorestframework-jsonapi>=5.0.0
django-filter>=23.5
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
pytest>=8.0.2
pytest-django>=4.8.0
flake8>=7.0.0
black>=24.2.0
isort>=5.13.2
drf-spectacular>=0.27.1 